ingestion_api/requirements.txt
# FastAPI Ingestion Service Dependencies

fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
confluent-kafka==2.3.0
pyyaml==6.0.1

